requirements.txt
autopep8==2.3.1
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
flake8==7.1.1
Flask==3.1.0
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyMySQL==1.1.1
PySocks==1.7.1
python-decouple==3.8
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.3.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
toml==0.10.2
tomli==2.2.1
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
